MVC - Refactor estructura de paquete
