--- app/src/main/java/ayds/newyork/songinfo/moredetails/fulllogic/model/repository/local/nyt/NytLocalStorage.kt
@@ -1,9 +1,0 @@
-package ayds.newyork.songinfo.moredetails.fulllogic.model.repository.local.nyt
-
-import ayds.newyork.songinfo.moredetails.fulllogic.model.entities.ArtistInfo
-
-interface NytLocalStorage {
-    fun saveArtist(artist: String, info: String)
-
-    fun getInfoByArtistName(artist: String): ArtistInfo
-}
