+++ app/src/main/java/ayds/newyork/songinfo/moredetails/view/MoreDetailsEvent.kt
@@ -1,4 +1,4 @@
-package ayds.newyork.songinfo.moredetails.fulllogic.view
+package ayds.newyork.songinfo.moredetails.view
 
 sealed class MoreDetailsUiEvent {
         object OpenArticleUrl : MoreDetailsUiEvent()
